Worked on the cli a bit
also added a util file that will act as our own standard library.
If we have access to the c stdlib without extra memory issues then
we should just use that and scrap the util module.

The cli probably doesn't function as of now, in fact I didn't even compile
it (also don't have the arduino ide), but I wanted to get a rough draft of
how I imagined it would work.